requirements.txt
discord.py==2.3.2
yt-dlp>=2024.1.1
PyNaCl>=1.5.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
ffmpeg-python>=0.2.0